pyproject.toml
[tool.poetry]
name = "pythoneda-realm-infrastructure-unveilingpartner"
version = "0.0.1a2"
description = "Infrastructure layer for UnveilingPartner's PythonEDA realm"
authors = ["rydnr <github@acm-sl.org>"]
readme = "README.md"
homepage = "https://github.com/pythoneda-realm-infrastructure/unveilingpartner"
packages = [{include = "pythonedarealminfrastructureunveilingpartner"}]

[tool.poetry.dependencies]
dbus-next = "^0.2.3"
grpcio = "^1.41.0"
oauth2 = "^1.9.0.post1"
python = "^3.8"
pythoneda-artifact-event-changes = "^0.0.1a1"
pythoneda-artifact-event-git-tagging = "^0.0.1a1"
pythoneda-artifact-event-infrastructure-changes = "^0.0.1a1"
pythoneda-artifact-event-infrastructure-git-tagging = "^0.0.1a1"
pythoneda-artifact-shared-changes = "^0.0.1a1"
pythoneda-base = "^0.0.1a18"
pythoneda-infrastructure-base = "^0.0.1a11"
pythoneda-realm-unveilingpartner = "^0.0.1a3"
pythoneda-shared-git = "^0.0.1a4"
requests = "^2.28.1"
requests-oauthlib = "^1.3.1"
unidiff = "^0.7.4"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
